Added the swap method to visualization
What I'm doing is largely inefficient and you will most likely see a
performance hit on the canvas redraw. We are redrawing each item in the
list rather than only swapping the two items for each swap. I need to
work out the grouping functionality of paperjs a little better to
improve on this.